Render Bug caused in viewDidAppear
I noticed that since viewDidAppear can be called more than once during a views life cycle we dont want to add a copy of the MessageInputBar to the view after is has already been correctly rendered. I just added a flag to account for this edge case. To see an example of the bug, present a UIAlertController in MessagesViewController and a copy of the MessageInputBar gets incorrectly added.